Complete the implementation of the GodotPayment plugin.
Move the remaining plugin components within the plugin source code.
